Update reading_progress field type to float and include it in document listing

+++ readwise/model.py
@@ -35,7 +35,7 @@ class Document(BaseModel):
     parent_id: Optional[str]
     saved_at: str
     last_moved_at: str
-    reading_progress: Optional[int]
+    reading_progress: Optional[float]
 
 
 class GetResponse(BaseModel):
